Interface/error corrections; Corrected logic in inherit and other optimizations; More tests
