+++ phalcon/Acl/AdapterInterface.zep
@@ -39,7 +39,7 @@ interface AdapterInterface
     /**
      * Adds access to components
      */
-    public function addComponentAccess(string componentName, accessList) -> void;
+    public function addComponentAccess(string componentName, accessList) -> bool;
 
     /**
      * Allow access to a role on a component
